requirements.txt
CherryPy==3.8.0
mutagen==1.28
python-vlc==1.1.2
ws4py==0.3.4
